refactor(ramp): changes folder structure and filename of AmountToBuy view
