+++ app/components/UI/Ramp/buy/Views/BuildQuote/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BuildQuote';
